+++ app/src/main/java/org/linphone/activities/main/chat/data/ChatMessageContentData.kt
@@ -482,5 +482,7 @@ interface OnContentClickedListener {
 
     fun onCallConference(address: String, subject: String?)
 
+    fun onShowReactionsList(chatMessage: ChatMessage)
+
     fun onError(messageId: Int)
 }